server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
minio
pydantic-settings
python-multipart
alembic
pyyaml

